requirements.txt
Django>=4.2,<5

djangorestframework>=3.14

django-environ>=0.9

# removed langchain dependency – using custom RAG components instead

faiss-cpu>=1.7.3

openai>=0.27.0

python-dotenv>=1.0  # optional if using .env

# cors headers
django-cors-headers>=3.15

# document parsing
pypdf>=3.12.0
python-docx>=0.8.12

# background/task queue
django-redis>=5.2
redis>=4.6
celery>=5.3

# tokenizers
tiktoken>=0.6

# ML/AI dependencies
# sentence-transformers>=2.2.0  # Optional: uncomment to use real sentence-transformers
# torch>=2.0.0  # Optional: required for sentence-transformers

# for development and testing
pytest-django>=4.5



